Refactor internal container impl.
