--- Softeq.XToolkit.WhiteLabel/Bootstrapper/Abstract/IConcreteRegistration.cs
@@ -1,10 +1,0 @@
-// Developed by Softeq Development Corporation
-// http://www.softeq.com
-
-namespace Softeq.XToolkit.WhiteLabel.Bootstrapper.Abstract
-{
-    public interface IConcreteRegistration
-    {
-        void PreserveExistingDefaults();
-    }
-}
